Compile translations for all projects on checkout
Instead of just compiling translations for stoq/kiwi/stoqdrivers, do that for
all projects on checkout dir that have a "po/pt_BR.po" file. Also, optimize it
to just compile the translations if the .po is newer than the .mo

We can do it on the script instead on the rc file since we are not using kiwi
anymore (msgfmt isn't on the custom PATH we set), so put it together with git
hooks processing.

Change-Id: I1163607804372a42c8629ea96eebadd6e0e204e0